requirements.txt
pandas
request
beautifulsoup4
lxml
matplotlib
circlify